requirements.txt
langchain
langchain-community
langgraph
langsmith
langchain-groq
python-dotenv
streamlit
faiss-cpu
pypdf
duckduckgo_search